+++ NEWS.md
@@ -405,7 +405,7 @@ Version 0.13 - 2020-07-04
 - NEW: `Spline.from_arc()`, create a new SPLINE entity from CIRCLE, ARC or ELLIPSE entity (constructor)
 - NEW: `Hatch.set_pattern_scale()` to set scaling of pattern definition
 - NEW: `Hatch.set_pattern_angle()` to set rotation angle of pattern definition
-- NEW: `Hatch.paths.polyline_to_edge_path()` convert polyline paths with bulge values to edge paths with lines and arcs
+- NEW: `Hatch.paths.polyline_to_edge_paths()` convert polyline paths with bulge values to edge paths with lines and arcs
 - NEW: `Hatch.paths.arc_edges_to_ellipse_edges()` convert arc edges to ellipse edges
 - NEW: `Hatch.paths.ellipse_edges_to_spline_edges()` convert ellipse edges to spline edges
 - NEW: `Hatch.paths.all_to_spline_edges()` convert all curves to approximated spline edges